+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Changed
+  * Updated column highlight first and last row detection for when a single row, and detecting empty and collapsed sections.
+
+* Added
+  * Added new row prop `isDecorative` to prevent column highlighting and interaction when consumers style rows as dividers.
+
 ## 2.35.0 - (March 25, 2022)
 
 * Added